tst: Move the `OutlineEffect` test to `HighlightCompoment.test.tsx`
