+++ packages/spatial/src/renderer/components/PostProcessingComponent.test.tsx
@@ -271,25 +271,6 @@ describe('PostProcessingComponent', () => {
       return destroyEngine()
     })
 
-    it('should add the OutlineEffect to the RendererComponent.effectComposer.EffectPass.effects list', async () => {
-      const effectKey = 'OutlineEffect'
-
-      // force nested reactors to run
-      const { rerender, unmount } = render(<></>)
-      await act(() => rerender(<></>))
-
-      // Check that the effect composer is setup
-      const effectComposer = getComponent(rootEntity, RendererComponent).effectComposer
-      assert.notEqual(Boolean(effectComposer), false, 'the effect composer is not setup correctly')
-
-      // Check that the effect pass has the the effect set
-      // @ts-ignore Allow access to the `effects` private field
-      const effects = effectComposer.EffectPass.effects
-      assert.equal(Boolean(effects.find((it) => it.name == effectKey)), true)
-
-      unmount()
-    })
-
     it('should add and remove effects correctly', async () => {
       const effectKey = 'NoiseEffect'
       noiseAddToEffectRegistry()
